requirements.txt
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==1.0.0
SQLAlchemy==2.0.7
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
